base done

+++ order_api/order_controller.py
@@ -1,5 +1,3 @@
-from os import abort
-
 from flask import Flask, jsonify, request
 from order_api.data.order_db import OrderDb
 
@@ -17,20 +15,25 @@ def device_output(*args):
         if arguments.get('id'):
             return jsonify(order_db.get_by_id(arguments.get('id')))
         else:
-            return order_db.get_all()
+            return jsonify(order_db.get_all())
 
     if request.method == 'PUT':
         if arguments.get('id'):
             request_json = request.json
-            return order_db.update(request_json['date'], request_json['customer_id'], request_json['order_lines'], arguments.get('id'))
+            return order_db.update(request_json['date'], request_json['customer_id'], request_json['order_lines'],
+                                   arguments.get('id'))
         else:
-            "Id needed"
+            'Id needed'
 
     if request.method == 'POST':
         request_json = request.json
         return order_db.insert(request_json['date'], request_json['customer_id'], request_json['order_lines'])
+
     if request.method == 'DELETE':
-        return order_db.delete(arguments.get('id'))
+        if arguments.get('id'):
+            return order_db.delete(arguments.get('id'))
+        else:
+            'Id needed'
 
 
 app.run(host='0.0.0.0')
